Move `sign-in.tsx` and `sign-out.tsx` to `auth` directory and update import paths
* **Move files**
  - Move `src/components/sign-in.tsx` to `src/components/auth/sign-in.tsx`
  - Move `src/components/sign-out.tsx` to `src/components/auth/sign-out.tsx`

* **Update import paths**
  - Update import path for `signIn` in `sign-in.tsx` to `@/lib/auth/auth`
  - Update import path for `signOut` in `sign-out.tsx` to `@/lib/auth/auth`
